temporary commit to (partially) proove the correctness of implementation using higher accuracy

+++ src/main.cpp
@@ -18,7 +18,7 @@ int main(int argc, char *argv[]) {
                 return 0;
             }
 
-            parser::ArcParser<float> parser;
+            parser::ArcParser<double> parser;
             auto structures_optional = parser.parse_file(arg);
             if (!structures_optional) {
                 std::cerr << "Error parsing file: " << arg << std::endl;
@@ -28,8 +28,8 @@ int main(int argc, char *argv[]) {
             for (const auto &structure : structures) {
                 uint64_t num_atoms = structure.atoms.size();
                 uint16_t *elements = (uint16_t *)malloc(sizeof(uint16_t) * num_atoms);
-                float *atoms = (float *)malloc(sizeof(float) * num_atoms * 3); // allocate memory for atom positions
-                float cell[3][3]; // allocate memory for cell parameters
+                double *atoms = (double *)malloc(sizeof(double) * num_atoms * 3); // allocate memory for atom positions
+                double cell[3][3]; // allocate memory for cell parameters
                 for (size_t i = 0; i < num_atoms; ++i) {
                     elements[i] = structure.atoms[i].element.atomic_number;
                     atoms[i * 3] = structure.atoms[i].position.x;
@@ -43,25 +43,25 @@ int main(int argc, char *argv[]) {
                     }
                 }
                 std::cout << "Computing dispersion energy for " << num_atoms << " atoms..." << std::endl;
-                float cutoff_radius = 60.0f; // cutoff radius in bohr
-                float CN_cutoff_radius = 40.0f; // cutoff radius in bohr
-                float energy;
-                float *force = (float *)malloc(sizeof(float) * num_atoms * 3); // allocate memory for force
-                float *stress = (float *)malloc(sizeof(float) * 9); // allocate memory for stress
+                double cutoff_radius = 60.0f; // cutoff radius in bohr
+                double CN_cutoff_radius = 40.0f; // cutoff radius in bohr
+                double energy;
+                double *force = (double *)malloc(sizeof(double) * num_atoms * 3); // allocate memory for force
+                double *stress = (double *)malloc(sizeof(double) * 9); // allocate memory for stress
                 // Start measuring execution time
                 auto start_time = std::chrono::high_resolution_clock::now();
-                compute_dispersion_energy((float (*)[3])atoms, elements, num_atoms, cell, cutoff_radius,
+                compute_dispersion_energy((double (*)[3])atoms, elements, num_atoms, cell, cutoff_radius,
                                           CN_cutoff_radius, ZERO_DAMPING, PBE, &energy, force, stress);
                 auto end_time = std::chrono::high_resolution_clock::now();
                 std::chrono::duration<double> elapsed_time = end_time - start_time;
                 std::cout << "Elapsed time: " << elapsed_time.count() << " seconds" << std::endl;
                 std::cout << std::fixed << std::setprecision(9) << "Energy: " << energy << " eV"
                           << std::endl; // convert to eV
-                float force_sum[3] = {0.0f, 0.0f, 0.0f};
+                double force_sum[3] = {0.0f, 0.0f, 0.0f};
                 for (size_t i = 0; i < num_atoms; ++i) {
-                    float force_x = force[0 + i * 3];
-                    float force_y = force[1 + i * 3];
-                    float force_z = force[2 + i * 3];
+                    double force_x = force[0 + i * 3];
+                    double force_y = force[1 + i * 3];
+                    double force_z = force[2 + i * 3];
                     force_sum[0] += force_x;
                     force_sum[1] += force_y;
                     force_sum[2] += force_z;
